+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Bug fixes
 
+- Fixed miss-calculation of signed rank in `ranktransform()` ( #87 ).
 - Fixed bug in `standardize()` for standard objects with non-standard class-attributes (like vectors of class `haven_labelled` or `vctrs_vctr`).
 
 # effectsize 0.3.1